LayoutTests/fast/webgpu/nocrash/fuzz-292945-expected.txt

CONSOLE MESSAGE: promises created
CONSOLE MESSAGE: the end
CONSOLE MESSAGE: fuzz-292945.html
CONSOLE MESSAGE: Pass
